Updates to the way that players and meeples are instantiated
Player and meeple instantiation has changed. PlayerScripts are now components of PhotonUser, so each new PhotonUser is a new player. Meeples are Instantiated by the players from a prefab. The old meeple object remain in the scene, but should be deleted in a future update.

I am not sure how well this works with multiple network players, so this will need to be tested!