Functional AI for placement of tiles & meeples
This AI is functional for placing tiles and meeples correctly. Contains a few workarounds to not have to rewrite the entire code, so it's not always the cleanest or most efficient solution. Still does not contain rewards for score, so it can't be trained to play well, only to play quickly and according to the rules.