Minor changes, code cleanup and "trained" AI.
Cleaned up some of the code to make it more readable. Also cleaned up the logging from the AI. It no longer uses error messages (some are still left as comments, for future bug identification) and leads each log with 'AI:' to indicate that it refers to an AI move. Fixed an issue that the AI could previously only rotate a tile to 3 of the 4 positions, as Chanon spotted. Also added skipping the initial "Start" button click event, so the AI can play on its own.
Regarding the 'trained' AI: This AI is NOT trained to be good at the game in any sense. It is merely a generated NN model so that the game with an AI that plays it can be started by just pressing play.