Working for single AI/Human
Fixed current development code to work for both a single AI or a single human-controlled player. When the game is launched, but before pressing 'Start Game' in the GUI, select the User1 object and change the boolean 'controlledByAI' to choose to play as human or AI.
This is not tested in multiplayer settings, and some parts of the code indicate that some changes may be needed for that.