Meeple placement logic changed & updated for rotation fix
Meeple placement now happens in one go, so the meeple is placed at the same time as AI makes the choice of where on the tile to place it, rather than first a location action, and then a confirm placement action.
Also fixed rotation to match the new update, where it happens in tile controller instead.