Working with RPC calls
Fixed the AI code to work via RPC calls all the way. Should function in multiplayer, assuming the multiplayer works as intended which it does not do right now (so this is not tested for more than one player).
Note that some of the code is not optimal due to the current code state. This is especially clear in the ConfirmPlacementAI-method in the section where the Phase is MeepleDrawn. This would need massive code rework to do in a different way.